requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
groq>=0.9.0
python-dotenv>=1.0.0
pytz>=2024.1
